backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
cryptography==41.0.7
pyotp==2.9.0
pytest==7.4.3
pylint==3.0.2
black==23.11.0
python-dotenv==1.0.0
marshmallow==3.20.1
flask-marshmallow==1.1.0
marshmallow-sqlalchemy==0.29.0
fido2==1.1.2
